Moved ValidationUtils.hpp to public API (#113)
