+++ include/powsybl/iidm/ValidationUtils.hpp
