+++ src/iidm/BusBreakerVoltageLevel.cpp
@@ -13,9 +13,9 @@
 
 #include <powsybl/iidm/Substation.hpp>
 #include <powsybl/iidm/Switch.hpp>
+#include <powsybl/iidm/ValidationUtils.hpp>
 
 #include "BusTerminal.hpp"
-#include "ValidationUtils.hpp"
 
 namespace powsybl {
