+++ src/iidm/Generator.cpp
@@ -7,9 +7,8 @@
 
 #include <powsybl/iidm/Generator.hpp>
 
+#include <powsybl/iidm/ValidationUtils.hpp>
 #include <powsybl/iidm/VariantManager.hpp>
-
-#include "ValidationUtils.hpp"
 
 namespace powsybl {
