+++ src/iidm/GeneratorAdder.cpp
@@ -8,11 +8,10 @@
 #include <powsybl/iidm/GeneratorAdder.hpp>
 
 #include <powsybl/iidm/Generator.hpp>
+#include <powsybl/iidm/ValidationUtils.hpp>
 #include <powsybl/iidm/VoltageLevel.hpp>
 #include <powsybl/stdcxx/make_unique.hpp>
 #include <powsybl/stdcxx/math.hpp>
-
-#include "ValidationUtils.hpp"
 
 namespace powsybl {
