+++ src/iidm/HvdcConverterStation.cpp
@@ -7,7 +7,7 @@
 
 #include <powsybl/iidm/HvdcConverterStation.hpp>
 
-#include "ValidationUtils.hpp"
+#include <powsybl/iidm/ValidationUtils.hpp>
 
 namespace powsybl {
