+++ src/iidm/LineAdder.cpp
@@ -5,16 +5,16 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
+#include <powsybl/iidm/LineAdder.hpp>
+
 #include <memory>
 
 #include <powsybl/iidm/Line.hpp>
-#include <powsybl/iidm/LineAdder.hpp>
 #include <powsybl/iidm/Network.hpp>
+#include <powsybl/iidm/ValidationUtils.hpp>
 #include <powsybl/iidm/VoltageLevel.hpp>
 #include <powsybl/stdcxx/make_unique.hpp>
 #include <powsybl/stdcxx/math.hpp>
-
-#include "ValidationUtils.hpp"
 
 namespace powsybl {
