+++ src/iidm/StaticVarCompensator.cpp
@@ -8,10 +8,9 @@
 #include <powsybl/iidm/StaticVarCompensator.hpp>
 
 #include <powsybl/iidm/Enum.hpp>
+#include <powsybl/iidm/ValidationUtils.hpp>
 #include <powsybl/iidm/VariantManager.hpp>
 #include <powsybl/stdcxx/format.hpp>
-
-#include "ValidationUtils.hpp"
 
 namespace powsybl {
