+++ src/iidm/TerminalBuilder.cpp
@@ -10,8 +10,7 @@
 #include <powsybl/iidm/Network.hpp>
 #include <powsybl/iidm/Validable.hpp>
 #include <powsybl/iidm/ValidationException.hpp>
-
-#include "ValidationUtils.hpp"
+#include <powsybl/iidm/ValidationUtils.hpp>
 
 namespace powsybl {
